combine all lambda code in one file
